requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
fpdf
python-pptx